requirements.txt
altgraph==0.17
cx-Freeze==6.1
numpy==1.18.2
PyInstaller==3.6
pyxel==1.3.1
PyYAML==5.3.1
